Remove temporary v2 router
