+++ api/router.go
@@ -46,7 +46,4 @@ func ConfigGinRouter(router *gin.Engine) {
 	configAuditRouter(router, daoWrapper)
 	configGinBookmarksRouter(router, daoWrapper)
 	configMaintenanceRouter(router, daoWrapper)
-
-	v2 := router.Group("/v2")
-	NewCoursesV2Router(v2, daoWrapper)
 }
